# backend/swagger_docs.yaml
openapi: 3.0.0
info:
  title: IDC设备管理系统 API
  version: 1.0.0
  description: |
    IDC设备管理系统后端API文档

    ## 认证说明
    除健康检查接口外，所有接口都需要在请求头中携带JWT Token：
    ```
    Authorization: Bearer <token>
    ```

    ## 通用响应格式
    ### 成功响应
    ```json
    {
      "success": true,
      "data": {...},
      "message": "操作成功"
    }
    ```

    ### 错误响应
    ```json
    {
      "success": false,
      "error": "错误类型",
      "message": "详细描述"
    }
    ```

    ## 分页参数
    列表接口支持 `page` 和 `pageSize` 参数，默认 page=1, pageSize=10
servers:
  - url: http://localhost:8000
    description: 开发环境服务器
security:
  - bearerAuth: []
...
paths:
...
  /api/rooms:
    get:
      summary: 获取机房列表
      tags: [rooms]
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
          description: 按名称搜索
        - name: status
          in: query
          schema:
            type: string
          description: 按状态筛选
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  roomId:
                    type: string
                  name:
                    type: string
                  location:
                    type: string
                  area:
                    type: number
                  status:
                    type: string
    post:
      summary: 创建机房
      tags: [rooms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roomId, name]
              properties:
                roomId:
                  type: string
                  description: 机房ID（唯一标识）
                name:
                  type: string
                  description: 机房名称
                location:
                  type: string
                  description: 机房位置
                area:
                  type: number
                  description: 面积(平方米)
                description:
                  type: string
                  description: 描述
                status:
                  type: string
                  enum: [active, inactive]
      responses:
        '201':
          description: 创建成功
  /api/rooms/{roomId}:
    get:
      summary: 获取单个机房
      tags: [rooms]
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
    put:
      summary: 更新机房
      tags: [rooms]
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                location:
                  type: string
                area:
                  type: number
                description:
                  type: string
                status:
                  type: string
      responses:
        '200':
          description: 更新成功
    delete:
      summary: 删除机房
      tags: [rooms]
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
...
  /api/consumables/logs:
    get:
      summary: 获取耗材操作日志
      tags: [consumables]
      parameters:
        - name: consumableId
          in: query
          schema:
            type: string
        - name: operationType
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
...
  /api/operation-logs:
    get:
      summary: 获取操作日志列表
      tags: [operation-logs]
      parameters:
        - name: userId
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
        - name: module
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: 成功
...
  /api/roles:
    get:
      summary: 获取角色列表
      tags: [roles]
      responses:
        '200':
          description: 成功
    post:
      summary: 创建角色
      tags: [roles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                roleName:
                  type: string
                description:
                  type: string
                permissions:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: 创建成功
  /api/roles/{roleId}:
    get:
      summary: 获取单个角色
      tags: [roles]
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
    put:
      summary: 更新角色
      tags: [roles]
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 更新成功
    delete:
      summary: 删除角色
      tags: [roles]
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 输入 JWT token